expense_tracker_en/requirements.txt
matplotlib
pandas
mysql-connector-python
seaborn